feat(channel rewards): toggle by sort tag
